Clean up errors, sync to parentBreadcrumbs, take out unnecessary app props
