+++ dashboards-observability/common/types/explorer.ts
@@ -104,13 +104,12 @@ export interface IExplorerProps {
   tabCreatedTypes?: any;
   searchBarConfigs?: any;
   appId?: string;
-  baseQuery?: string;
   addVisualizationToPanel?: any;
   startTime?: string;
   endTime?: string;
   setStartTime?: any;
   setEndTime?: any;
-  appBaseQuery: string;
+  appBaseQuery?: string;
 }
 
 export interface SavedQuery {
