+++ dashboards-observability/public/components/explorer/explorer.tsx
@@ -42,7 +42,6 @@ import {
   SELECTED_TIMESTAMP,
   AVAILABLE_FIELDS,
   TIME_INTERVAL_OPTIONS,
-  HAS_SAVED_TIMESTAMP,
   SAVED_QUERY,
   SAVED_VISUALIZATION,
   SAVED_OBJECT_ID,
