+++ dashboards-observability/public/components/index.tsx
@@ -10,9 +10,9 @@ import { AppPluginStartDependencies } from '../types';
 import { App } from './app';
 
 export const Observability = (
-  CoreStart: CoreStart,
+  CoreStartProp: CoreStart,
   DepsStart: AppPluginStartDependencies,
-  AppMountParameters: AppMountParameters,
+  AppMountParametersProp: AppMountParameters,
   pplService: any,
   dslService: any,
   savedObjects: any,
@@ -20,15 +20,15 @@ export const Observability = (
 ) => {
   ReactDOM.render(
     <App
-      CoreStart={CoreStart}
+      CoreStartProp={CoreStartProp}
       DepsStart={DepsStart}
       pplService={pplService}
       dslService={dslService}
       savedObjects={savedObjects}
       timestampUtils={timestampUtils}
     />,
-    AppMountParameters.element
+    AppMountParametersProp.element
   );
 
-  return () => ReactDOM.unmountComponentAtNode(AppMountParameters.element);
+  return () => ReactDOM.unmountComponentAtNode(AppMountParametersProp.element);
 };
